[ISSUE-146] refactor sort logic to make it more generic
