+++ app/views/admin/volunteers/_volunteers_table.html.erb
@@ -1,4 +1,4 @@
-<table class="table" id="volunteers">
+<table class="table" id="sortable">
   <thead class="table-gray">
     <tr>
       <th
@@ -36,7 +36,7 @@
     </tr>
   </thead>
   <tbody>
-    <% volunteers.each do |volunteer| %>
+    <% records.each do |volunteer| %>
       <tr scope="row">
         <td>
           <%= volunteer.first_name %>
