+++ app/views/admin/volunteers/index.html.erb
@@ -22,7 +22,7 @@
   </div>
 </div>
 
-<%= render "volunteers_table", volunteers: @volunteers %>
+<%= render "volunteers_table", records: @volunteers %>
 
 <%= render(
         layout: "admin/shared/csv_import_modal",
